+++ webpack.config.js
@@ -22,7 +22,10 @@ const InlineConstantExportsPlugin = require( '@automattic/webpack-inline-constan
 const Minify = require( '@automattic/calypso-build/webpack/minify' );
 const SassConfig = require( '@automattic/calypso-build/webpack/sass' );
 const TranspileConfig = require( '@automattic/calypso-build/webpack/transpile' );
-const { cssNameFromFilename } = require( '@automattic/calypso-build/webpack/util' );
+const {
+	cssNameFromFilename,
+	IncrementalProgressPlugin,
+} = require( '@automattic/calypso-build/webpack/util' );
 const ExtensiveLodashReplacementPlugin = require( '@automattic/webpack-extensive-lodash-replacement-plugin' );
 
 /**
@@ -43,6 +46,7 @@ const shouldMinify =
 	process.env.MINIFY_JS === 'true' ||
 	( process.env.MINIFY_JS !== 'false' && bundleEnv === 'production' && calypsoEnv !== 'desktop' );
 const shouldEmitStats = process.env.EMIT_STATS && process.env.EMIT_STATS !== 'false';
+const shouldShowProgress = process.env.PROGRESS && process.env.PROGRESS !== 'false';
 const shouldEmitStatsWithReasons = process.env.EMIT_STATS === 'withreasons';
 const shouldCheckForCycles = process.env.CHECK_CYCLES === 'true';
 const isCalypsoClient = process.env.BROWSERSLIST_ENV !== 'server';
@@ -54,54 +58,6 @@ const extraPath = browserslistEnv === 'defaults' ? 'fallback' : browserslistEnv;
 
 if ( ! process.env.BROWSERSLIST_ENV ) {
 	process.env.BROWSERSLIST_ENV = browserslistEnv;
-}
-
-/*
- * Create reporter for ProgressPlugin (used with EMIT_STATS)
- */
-function createProgressHandler() {
-	const startTime = Date.now();
-	let lastShownBuildingMessageTime = null;
-	let lastUnshownBuildingMessage = null;
-
-	return ( percentage, msg, ...details ) => {
-		const nowTime = Date.now();
-		const timeString = ( ( nowTime - startTime ) / 1000 ).toFixed( 1 ) + 's';
-		const percentageString = `${ Math.floor( percentage * 100 ) }%`;
-		const detailsString = details
-			.map( detail => {
-				if ( ! detail ) {
-					return '';
-				}
-				if ( detail.length > 40 ) {
-					return `…${ detail.substr( detail.length - 39 ) }`;
-				}
-				return detail;
-			} )
-			.join( ' ' );
-		const message = `${ timeString } ${ percentageString } ${ msg } ${ detailsString }`;
-
-		// There are plenty of 'building' messages that make the log too long for CircleCI web UI.
-		// Let's throttle the 'building' messages to one per second, while always showing the last one.
-		if ( msg === 'building' ) {
-			if ( lastShownBuildingMessageTime && nowTime - lastShownBuildingMessageTime < 1000 ) {
-				// less than 1000ms since last message: ignore, but store for case it's the last one
-				lastUnshownBuildingMessage = message;
-				return;
-			}
-
-			// the message will be shown and its time recorded
-			lastShownBuildingMessageTime = nowTime;
-			lastUnshownBuildingMessage = null;
-		} else if ( lastUnshownBuildingMessage ) {
-			// The last 'building' message should always be shown, no matter the timing.
-			// We do that on the first non-'building' message.
-			console.log( lastUnshownBuildingMessage ); // eslint-disable-line no-console
-			lastUnshownBuildingMessage = null;
-		}
-
-		console.log( message ); // eslint-disable-line no-console
-	};
 }
 
 const nodeModulesToTranspile = [
@@ -329,7 +285,7 @@ const webpackConfig = {
 					chunkGroups: true,
 				},
 			} ),
-		shouldEmitStats && new webpack.ProgressPlugin( createProgressHandler() ),
+		shouldShowProgress && new IncrementalProgressPlugin(),
 		new MomentTimezoneDataPlugin( {
 			startYear: 2000,
 			cacheDir: path.join(